+++ src/cli.ts
@@ -7,9 +7,16 @@ import { showUsage } from './usage';
 
 const [tsconfig, ...tsFiles] = process.argv.slice(2);
 
+// eslint style exit code:
+enum ExitCode {
+  NoUnusedExportsFound = 0,
+  UnusedExportsFound = 1,
+  BadArgsOrException = 2,
+}
+
 if (!hasValidArgs()) {
   showUsage();
-  process.exit(-1);
+  process.exit(ExitCode.BadArgsOrException);
 }
 
 try {
@@ -59,7 +66,13 @@ try {
   if (options && options.exitWithCount) {
     process.exit(files.length);
   }
+
+  process.exit(
+    files.length === 0
+      ? ExitCode.NoUnusedExportsFound
+      : ExitCode.UnusedExportsFound,
+  );
 } catch (e) {
   console.error(e);
-  process.exit(-1);
+  process.exit(ExitCode.BadArgsOrException);
 }